requirements.txt
apispec==6.3.0
asgiref==3.7.2
Django==4.2.3
django-serializer==1.2.1
djangorestframework==3.14.0
marshmallow==3.19.0
packaging==23.1
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
